+++ packages/l10n_countries/example/lib/main.dart
@@ -1,7 +1,9 @@
 // ignore_for_file: avoid_print, it's just an example.
 import "package:l10n_countries/l10n_countries.dart";
 
-void main() {
+void main(List<String> args) {
+  final mainLocale = args.firstOrNull ?? "fr";
+
   /// Provide a custom locale with a new language code (SCT -> Scotland).
   final mapper = CountriesLocaleMapper(
     other: {
@@ -12,14 +14,14 @@ void main() {
 
   final localized = mapper.localize(
     const {"USA", "POL", "SCT"}, // Country codes: USA, Poland, Scotland.
-    mainLocale: "fr", // Aim to get French L10N for country names first.
-    fallbackLocale: "en_US", // Fallback to English if French L10N is missing.
+    mainLocale: mainLocale, // Aim to get specific L10N for country names 1st.
+    fallbackLocale: "en_US", // Fallback to English if specific L10N is missing.
     formatter:
         (localeKey, l10n) => // Optional formatter, uppercase for English L10N.
             localeKey.locale.startsWith("en") ? l10n.toUpperCase() : l10n,
   );
 
-  print("Names count ${localized.length}"); // 9 - no French entry for Scotland.
+  print("Names count ${localized.length}"); // 9 - no main locale entry for SCT.
   /// Print out the localized names.
   localized.forEach(
     (country, l10n) => print(