+++ packages/l10n_languages/example/lib/main.dart
@@ -1,7 +1,9 @@
 // ignore_for_file: avoid_print, it's just an example.
 import "package:l10n_languages/l10n_languages.dart";
 
-void main() {
+void main(List<String> args) {
+  final mainLocale = args.firstOrNull ?? "fr";
+
   /// Provide a custom locale with a new language code (SCO -> Scots language).
   final mapper = LanguagesLocaleMapper(
     other: {
@@ -12,14 +14,14 @@ void main() {
 
   final localized = mapper.localize(
     const {"ENG", "POL", "SCO"}, // Language codes: English, Polish, Scots.
-    mainLocale: "fr", // Aim to get French L10N for language names first.
-    fallbackLocale: "en_US", // Fallback to English if French L10N is missing.
+    mainLocale: mainLocale, // Aim to get specific L10N for language names 1st.
+    fallbackLocale: "en_US", // Fallback to English if specific L10N is missing.
     formatter:
         (localeKey, l10n) => // Optional formatter, uppercase for English L10N.
             localeKey.locale.startsWith("en") ? l10n.toUpperCase() : l10n,
   );
 
-  print("Names count ${localized.length}"); // 5 - no French entry for Scots.
+  print("Names count ${localized.length}"); // 5 - no specific entry for Scots.
   /// Print out the localized names.
   localized.forEach(
     (language, l10n) => print(