+++ packages/world_countries/example/lib/main.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: avoid-late-keyword
+// ignore_for_file: avoid-late-keyword, avoid-disposing-late-fields
 
 import "package:flutter/material.dart";
 import "package:flutter_localizations/flutter_localizations.dart";